requirements.txt
playwright
pandas
openpyxl
python-dotenv
google-generativeai
deep-translator
